Updating to use backup to drive deletes.
Initial version of this script did fancy navigation to find
all-the-things.

Twitter's UI doesn't actually show you all-the-things, and will
start to show you less as you delete a lot of things.

A twitter data backup, however, has a full list of tweet IDs.

This script now takes that to examine and delete all possible
tweets.